Use pure CSS instead of JS for the blog BG color
