+++ src/pages/index.tsx
@@ -1,15 +1,9 @@
-import React, { useEffect, useRef, useState } from 'react';
 import Head from 'next/head';
 import Image from 'next/image';
-import blogCow from '../../public/images/Blog-cow.jpg';
 import heroBackground from '../../public/images/VH-hero-bg.jpg';
 import heroTagline from '../../public/images/VH-hero-tagline.png';
 import pixelHeart from '../../public/images/VH_PixelHeart.png';
-import {
-  DarkButton,
-  GreenButton,
-  LightButton,
-} from '../components/decoration/buttons';
+import { DarkButton, LightButton } from '../components/decoration/buttons';
 import Hero from '../components/decoration/hero';
 import SquareField from '../components/decoration/squares';
 import { GrantsCallToAction } from '../components/layout/grants/index';
@@ -19,7 +13,8 @@ import FeaturedProject from '../components/layout/index/featuredProject';
 import type { IProject } from '../types/generated/contentful';
 import type { GetStaticProps } from 'next';
 import { getFeaturedProjects } from '../lib/cms/helpers';
-import ImageContainer from '../components/decoration/imageContainer';
+import {} from '../components/decoration/textBlocks';
+import LastBlogEntries from '../components/layout/index/lastBlogEntries';
 
 const HERO_DECORATION_SQUARES = [
   { color: 'green', size: 32, left: 0, bottom: 0 },
@@ -48,10 +43,6 @@ const BLOG_DECORATION_SQUARES = [
   { color: 'gray-lighter', size: 16, right: 0, top: 0 },
 ];
 
-const BLOG_INNER_DECORATION_SQUARES = [
-  { color: 'gray-lighter', size: 16, right: 0, bottom: 0 },
-];
-
 interface HomeProps {
   featuredProjects: IProject[];
 }
@@ -63,38 +54,6 @@ export const getStaticProps: GetStaticProps = async () => {
 };
 
 const Home: React.FC<HomeProps> = ({ featuredProjects }) => {
-  const [grayAreaHeight, setGrayAreaHeight] = useState(0);
-  const blogCardText = useRef<HTMLDivElement>(null);
-  const blogCardButton = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    // Lines up gray area at bottom of page with the bottom of the blog images
-    if (
-      grayAreaHeight === 0 &&
-      blogCardText.current &&
-      blogCardButton.current
-    ) {
-      setGrayAreaHeight(
-        blogCardText.current.clientHeight + blogCardButton.current.clientHeight
-      );
-
-      const handleResize = () => {
-        if (blogCardText.current && blogCardButton.current) {
-          setGrayAreaHeight(
-            blogCardText.current.clientHeight +
-              blogCardButton.current.clientHeight
-          );
-        }
-      };
-
-      window.addEventListener('resize', handleResize);
-
-      return () => {
-        window.removeEventListener('resize', handleResize);
-      };
-    }
-  }, []);
-
   return (
     <>
       <Head>
@@ -201,106 +160,7 @@ const Home: React.FC<HomeProps> = ({ featuredProjects }) => {
         squares={BLOG_DECORATION_SQUARES}
         className="hidden md:block"
       />
-      <div className="content-center mx-auto md:w-1/2 drop-shadow-2xl text-2xl pt-16">
-        <p className="text-grey-dark pb-5">
-          <span className="font-italic font-semibold text-3xl">On the </span>
-          <b className="text-5xl font-mono">BLOG</b>
-        </p>
-        <p className="pb-16">
-          We regularly post project updates, announcements, interviews, and
-          other fun stuff here! Thanks for reading!
-        </p>
-      </div>
-      <div className="flex md:flex-row flex-col md:space-x-4 md:space-y-0 space-y-4 lg:px-32 px-4">
-        <div>
-          <div className="overflow-hidden w-full">
-            <ImageContainer>
-              <Image
-                src={blogCow.src}
-                width={blogCow.width}
-                height={blogCow.height}
-                className="w-full bg-cover"
-                alt="Compassion, Creativity, Code"
-              />
-            </ImageContainer>
-          </div>
-          <div className="overflow-hidden w-full">
-            <div className="px-8 pt-6 pb-6 content-center mx-auto bg-white">
-              <p className="font-semibold text-2xl">
-                Developers! Join Our New Open Source Community - VH: Playground
-              </p>
-            </div>
-            <GreenButton href="/">Read More</GreenButton>
-          </div>
-        </div>
-        <div>
-          <div className="overflow-hidden w-full">
-            <ImageContainer>
-              <Image
-                src={blogCow.src}
-                width={blogCow.width}
-                height={blogCow.height}
-                className="w-full bg-cover"
-                alt="Compassion, Creativity, Code"
-              />
-            </ImageContainer>
-          </div>
-          <div className="overflow-hidden w-full">
-            <div className="px-8 pt-6 pb-6 content-center mx-auto bg-white">
-              <p className="font-semibold text-2xl">
-                Developers! Join Our New Open Source Community - VH: Playground
-              </p>
-            </div>
-            <GreenButton href="/">Read More</GreenButton>
-          </div>
-        </div>
-        <div>
-          <div className="overflow-hidden w-full">
-            <ImageContainer>
-              <Image
-                src={blogCow.src}
-                width={blogCow.width}
-                height={blogCow.height}
-                className="w-full bg-cover"
-                alt="Compassion, Creativity, Code"
-              />
-            </ImageContainer>
-          </div>
-          <div className="overflow-hidden w-full">
-            <div
-              className="px-8 pt-6 pb-6 content-center mx-auto bg-white"
-              ref={blogCardText}
-            >
-              <p className="font-semibold text-2xl">
-                Developers! Join Our New Open Source Community - VH: Playground
-              </p>
-            </div>
-            <div ref={blogCardButton}>
-              <GreenButton href="/">Read More</GreenButton>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div className="relative" style={{ bottom: grayAreaHeight }}>
-        <SquareField
-          squares={BLOG_INNER_DECORATION_SQUARES}
-          className="hidden lg:block"
-        />
-      </div>
-      <div className="bg-grey relative">
-        <div
-          className="w-full bg-grey absolute"
-          style={{ height: grayAreaHeight, top: -grayAreaHeight, zIndex: -1 }}
-        />
-        <div className="relative mx-auto pt-10 md:w-1/3 pb-16 sm:px-24 px-20">
-          <LightButton
-            href="/projects"
-            className="font-mono text-sm font-semibold"
-          >
-            See All Posts
-          </LightButton>
-        </div>
-      </div>
+      <LastBlogEntries />
       <SquareField
         squares={PROJECT_DECORATION_SQUARES}
         className="hidden md:block"
